Revert "feat: support for solana contracts and releases"
