+++ src/evm/releases/resolvers.ts
@@ -1,8 +1,8 @@
 import { chainsQueries } from "@src/evm/chains/queries";
-import { getNestedValues } from "@src/helpers";
-import { createContractMapper, createStandardDeploymentResolver } from "@src/internal/factories/resolver";
-import type { AliasMap } from "@src/shared/types";
+import { getNestedValues } from "@src/evm/releases/helpers";
+import { getContractExplorerURL } from "@src/helpers";
 import type { Sablier } from "@src/types";
+import _ from "lodash";
 
 /* -------------------------------------------------------------------------- */
 /*                                   TYPES                                    */
@@ -12,7 +12,7 @@ type DeploymentBaseParams = {
   protocol: Sablier.EVM.Protocol;
   version: Sablier.EVM.Version;
   chainId: number;
-  aliasMap: AliasMap;
+  aliasMap: Sablier.EVM.AliasMap;
 };
 
 type DeploymentLockupV1Params = DeploymentBaseParams & {
@@ -29,49 +29,6 @@ type DeploymentStandardParams = DeploymentBaseParams & {
 type ReleaseParams<T> = Omit<T, "kind" | "contractNames">;
 
 /* -------------------------------------------------------------------------- */
-/*                             PLATFORM SETUP                                 */
-/* -------------------------------------------------------------------------- */
-
-const contractMapper = createContractMapper<
-  Sablier.EVM.Contract,
-  Sablier.EVM.Protocol,
-  Sablier.EVM.Version,
-  Sablier.EVM.Address
->(chainsQueries);
-
-const standardDeploymentResolver = createStandardDeploymentResolver<
-  Sablier.EVM.Deployment,
-  Sablier.EVM.Contract,
-  Sablier.EVM.Protocol,
-  Sablier.EVM.Version,
-  Sablier.EVM.Address
->(contractMapper, "contracts");
-
-/* -------------------------------------------------------------------------- */
-/*                           EVM-SPECIFIC RESOLVERS                           */
-/* -------------------------------------------------------------------------- */
-
-/**
- * Creates a LockupV1 deployment with separate core and periphery contracts
- */
-function createLockupV1Deployment(params: DeploymentLockupV1Params): Sablier.EVM.Deployment.LockupV1 {
-  const { contractMap, ...baseParams } = params;
-
-  // Create standard deployment with merged contracts
-  const mergedContracts = { ...contractMap.core, ...contractMap.periphery };
-  const deployment = standardDeploymentResolver({
-    ...baseParams,
-    contractMap: mergedContracts,
-  }) as Sablier.EVM.Deployment.LockupV1;
-
-  // Add separated core and periphery contracts
-  deployment.core = contractMapper(contractMap.core, baseParams);
-  deployment.periphery = contractMapper(contractMap.periphery, baseParams);
-
-  return deployment;
-}
-
-/* -------------------------------------------------------------------------- */
 /*                                 RESOLVERS                                  */
 /* -------------------------------------------------------------------------- */
 
@@ -81,14 +38,33 @@ export const resolvers = {
      * Creates a LockupV1 deployment with separate core and periphery contracts
      */
     lockupV1: (params: DeploymentLockupV1Params): Sablier.EVM.Deployment.LockupV1 => {
-      return createLockupV1Deployment(params);
+      const { contractMap, ...baseParams } = params;
+
+      // Create standard deployment with merged contracts
+      const mergedContracts = { ...contractMap.core, ...contractMap.periphery };
+      const deployment = resolvers.deployment.standard({
+        ...baseParams,
+        contractMap: mergedContracts,
+      }) as Sablier.EVM.Deployment.LockupV1;
+
+      // Add separated core and periphery contracts
+      deployment.core = mapContractsToDeployment(contractMap.core, baseParams);
+      deployment.periphery = mapContractsToDeployment(contractMap.periphery, baseParams);
+
+      return deployment;
     },
 
     /**
      * Creates a standard deployment with all contracts in a single array
      */
     standard: (params: DeploymentStandardParams): Sablier.EVM.Deployment => {
-      return standardDeploymentResolver(params);
+      const { contractMap, ...baseParams } = params;
+      const contracts = mapContractsToDeployment(contractMap, baseParams);
+
+      return {
+        chainId: baseParams.chainId,
+        contracts,
+      };
     },
   },
 
@@ -116,3 +92,33 @@ export const resolvers = {
     },
   },
 };
+
+/* -------------------------------------------------------------------------- */
+/*                                  HELPERS                                   */
+/* -------------------------------------------------------------------------- */
+
+/**
+ * Converts a contract map to an array of deployment contracts
+ */
+function mapContractsToDeployment(
+  contractMap: Sablier.EVM.ContractMap,
+  params: Pick<DeploymentBaseParams, "chainId" | "protocol" | "version" | "aliasMap">,
+): Sablier.EVM.Contract[] {
+  const { chainId, protocol, version, aliasMap } = params;
+  const chain = chainsQueries.getOrThrow(chainId);
+
+  return _.entries(contractMap).map(([name, addressOrTuple]) => {
+    const [address, blockNumber] = Array.isArray(addressOrTuple) ? addressOrTuple : [addressOrTuple];
+
+    return {
+      address,
+      alias: aliasMap[name],
+      block: blockNumber,
+      chainId,
+      explorerURL: getContractExplorerURL(chain.blockExplorers.default.url, address),
+      name,
+      protocol,
+      version,
+    };
+  });
+}
