+++ src/helpers.ts
@@ -1,4 +1,3 @@
-import _ from "lodash";
 import type { Sablier } from "./types";
 
 /**
@@ -59,13 +58,4 @@ export function isVersionAfter(version: Sablier.EVM.Version, after: Sablier.EVM.
 
 export function sortChains<T extends { name: string }>(chains: T[]): T[] {
   return chains.sort((a, b) => a.name.localeCompare(b.name));
-}
-
-export function getNestedValues<T extends Record<string, unknown>>(obj: T): string[] {
-  return _.flatMap(obj, (value) => {
-    if (_.isObject(value) && !_.isArray(value)) {
-      return getNestedValues(value as Record<string, unknown>);
-    }
-    return _.isString(value) ? value : [];
-  });
 }
