+++ src/sablier.ts
@@ -28,8 +28,6 @@ import { chainsQueries as evmChainsQueries } from "./evm/chains/queries";
 import { contractsQueries as evmContractsQueries } from "./evm/contracts/queries";
 import { releasesQueries as evmReleasesQueries } from "./evm/releases/queries";
 import { chainsQueries as solanaChainsQueries } from "./solana/chains/queries";
-import { contractsQueries as solanaContractsQueries } from "./solana/contracts/queries";
-import { releasesQueries as solanaReleasesQueries } from "./solana/releases/queries";
 
 /**
  * Has to be defined here to avoid circular dependencies.
@@ -49,21 +47,6 @@ const evmDeploymentsQueries = {
   },
 };
 
-const solanaDeploymentsQueries = {
-  /**
-   * Get many deployments.
-   * - default            ⇒ all across all releases
-   * - release            ⇒ that release's deployments
-   */
-  get: (opts: { chainId: number; release: Sablier.Solana.Release }): Sablier.Solana.Deployment | undefined => {
-    const { release, chainId } = opts || {};
-    return _.find(release.deployments, { chainId });
-  },
-  getAll: (): Sablier.Solana.Deployment[] => {
-    return _.flatMap(solanaReleasesQueries.getAll(), (r) => r.deployments);
-  },
-};
-
 const evm = {
   chains: evmChainsQueries,
   contracts: evmContractsQueries,
@@ -73,9 +56,6 @@ const evm = {
 
 const solana = {
   chains: solanaChainsQueries,
-  contracts: solanaContractsQueries,
-  deployments: solanaDeploymentsQueries,
-  releases: solanaReleasesQueries,
 };
 
 export const sablier = {
