--- src/solana/contracts/index.ts
@@ -1,7 +1,0 @@
-import { catalog } from "./catalog";
-import { names } from "./names";
-
-export const contracts = {
-  catalog,
-  names,
-};
