+++ src/solana/idl/airdrops/v0.1/SablierMerkleInstant/errors.ts
