+++ src/solana/idl/airdrops/v0.1/SablierMerkleInstant/structs.ts
