+++ src/solana/idl/lockup/v0.1/SablierLockupLinear/errors.ts
