+++ src/solana/index.ts
@@ -1,5 +1,3 @@
 export { chains } from "./chains";
-export * from "./contracts";
 export * from "./enums";
-export * from "./releases";
 export * from "./types";
