--- src/solana/releases/airdrops/v0.1/manifest.ts
@@ -1,3 +1,0 @@
-const manifest = { SABLIER_MERKLE_INSTANT: "SablierMerkleInstant" };
-
-export default manifest;
