--- src/solana/releases/data.ts
@@ -1,4 +1,0 @@
-import { airdrops } from "./airdrops";
-import { lockup } from "./lockup";
-
-export const releases = { airdrops, lockup };
