--- src/solana/releases/lockup/v0.1/aliases.ts
@@ -1,5 +1,0 @@
-import manifest from "./manifest";
-
-const aliases = { [manifest.SABLIER_LOCKUP_LINEAR]: "LL" };
-
-export default aliases;
