+++ src/solana/types.ts
@@ -1,4 +1,4 @@
-import type { AliasMap, Repository, Shared } from "@src/shared/types";
+import type { Shared } from "@src/shared/types";
 import type * as enums from "./enums";
 
 export namespace Solana {
@@ -22,60 +22,7 @@ export namespace Solana {
     };
   };
 
-  /**
-   * The base contract type for Solana.
-   */
-  export type Program = Shared.Contract<Address, Protocol, Version>;
-
-  export type ProgramMap = Shared.ContractMap<Address>;
-  export type ProgramCatalog = {
-    [protocol in Protocol]: {
-      [chainId: number]: {
-        [address: Address]: Program;
-      };
-    };
-  };
-
-  /**
-   * A collection of contracts deployed on a single chain.
-   */
-  export type Deployment = {
-    chainId: number;
-    programs: Program[];
-  };
-
-  export type IdlMap = { [contractName: string]: object };
-
-  /**
-   * Contract names for a given protocol and version.
-   */
-  export type Manifest = Shared.Manifest;
-
   export type Protocol = `${enums.Protocol}` | enums.Protocol;
-
-  /**
-   * A collection of deployments for a given protocol and version.
-   */
-  export type Release = {
-    /** A map of contract names to their IDL (Interface Definition Language). */
-    idl: IdlMap;
-    /** A map of contract names to their aliases, used in the Sablier Interface and indexers. */
-    aliases?: AliasMap;
-    /** An array of contract names. */
-    programNames: string[];
-    /** List of deployments across different chains. */
-    deployments: Deployment[];
-    /** Whether this is the latest release for this protocol. */
-    isLatest: boolean;
-    /** The contract name manifest. */
-    manifest: Manifest;
-    /** The Sablier protocol released, e.g. `lockup`. */
-    protocol: Protocol;
-    /** Repository information for the release. */
-    repository?: Repository;
-    /** The version of the release, e.g., `v1.0`. */
-    version: Version;
-  };
 
   export namespace Version {
     export type Airdrops = `${enums.Version.Airdrops}` | enums.Version.Airdrops;
